+++ lib/jahuty/client.rb
@@ -7,28 +7,23 @@ module Jahuty
   class Client
     def initialize(api_key:, cache: nil, expires_in: nil)
       @api_key    = api_key
-      @cache      = cache || ::MiniCache::Store.new
+      @cache      = Cache::Facade.new(cache || ::MiniCache::Store.new)
       @expires_in = expires_in
-      @services   = Service::Factory.new(client: self)
+      @services   = {}
     end
 
     # Allows services to be accessed as properties (e.g., jahuty.snippets).
     def method_missing(name, *args, &block)
-      if args.empty? && @services.respond_to?(name)
-        @services.send(name)
+      if args.empty? && name == :snippets
+        unless @services.key?(name)
+          @services[name] = Service::Snippet.new(
+            client: self, cache: @cache, expires_in: @expires_in
+          )
+        end
+        @services[name]
       else
         super
       end
-    end
-
-    def fetch(action, expires_in: nil)
-      @manager ||= Cache::Manager.new(
-        client: self,
-        cache: @cache,
-        expires_in: expires_in || @expires_in
-      )
-
-      @manager.fetch(action)
     end
 
     def request(action)
@@ -40,17 +35,17 @@ module Jahuty
 
       response = @client.send(request)
 
-      @resources ||= Resource::Factory.new
+      @responses ||= Response::Handler.new
 
-      resource = @resources.call(action, response)
+      result = @responses.call(action, response)
 
-      raise Exception::Error.new(resource), 'API responded with a problem' if resource.is_a?(Resource::Problem)
+      raise Exception::Error.new(result), 'API problem' if result.is_a?(Resource::Problem)
 
-      resource
+      result
     end
 
     def respond_to_missing?(name, include_private = false)
-      @services.respond_to?(name, include_private) || super
+      name == :snippets || super
     end
   end
 end