+++ spec/jahuty/resource/render_spec.rb
@@ -3,13 +3,18 @@
 module Jahuty
   module Resource
     RSpec.describe Render do
-      subject(:render) { described_class.new(content: 'foo') }
+      subject(:render) { described_class.new(content: content, snippet_id: snippet_id) }
 
-      let(:content) { 'foo' }
+      let(:snippet_id) { 1 }
+      let(:content)    { 'foo' }
 
       describe '#initialize' do
         it 'sets content' do
           expect(render.content).to eq(content)
+        end
+
+        it 'sets snippet_id' do
+          expect(render.snippet_id).to eq(snippet_id)
         end
       end
 
@@ -30,10 +35,16 @@ module Jahuty
       end
 
       describe '::from' do
-        let(:data) { { content: 'foo' } }
+        let(:data) { { content: 'foo', snippet_id: 1 } }
 
         it 'raises error if :content key does not exist' do
           data.delete(:content)
+
+          expect { described_class.from(data) }.to raise_error(ArgumentError)
+        end
+
+        it 'raises error if :snippet_id key does not exist' do
+          data.delete(:snippet_id)
 
           expect { described_class.from(data) }.to raise_error(ArgumentError)
         end
@@ -42,6 +53,10 @@ module Jahuty
           data[:foo] = 'bar'
 
           expect { described_class.from(data) }.not_to raise_error
+        end
+
+        it 'returns render' do
+          expect(described_class.from(data)).to be_instance_of(described_class)
         end
       end
     end